@normanmaurer 2017-09-21
@carl-mastrangelo can you please rebase ?