@carl-mastrangelo 2017-10-17
@normanmaurer and @Scottmitch   

friendly ping.  I was hoping this could see some pre release usage.  If there are any other concerns, would there be another reviewer to look at this?