@Scottmitch 2017-11-11
sorry for delay ... I took a look and discussed offline with @carl-mastrangelo